requirements.txt
Mako==1.0.6
protobuf==3.11.0
mock==2.0.0
astroid==1.5.3
isort==4.2.15
pylint==1.7.2
nose==1.3.7
parameterized==0.6.1
cantools==32.19.0
paho-mqtt==1.5.0
python-can==3.3.3